adding some battle features
